+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
 - [FP-1516 - Changed the IO/Config Port icons to be more coherent](https://movai.atlassian.net/browse/FP-1516)
 - [FP-1612 - Only reset ports when need](https://movai.atlassian.net/browse/FP-1612)
+- [FP-1625 - Fixed a few issues regarding Node Editor](https://movai.atlassian.net/browse/FP-1625)
 
 **Home Tab**